chore(deepseek): split runtime module
